Move the update of the OIDC metadata after all Configure<JwtBearerOptions> are done (that is in the JwtBearerOptionsMerger, which is a PostConfigure
Fixed a flaky test

Added code to initialize Playwright
