+++ src/Microsoft.Identity.Web.TokenAcquisition/AspNetCore/JwtBearerOptionsMerger.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System.Linq;
+using System;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.Extensions.Options;
 
@@ -23,7 +25,22 @@ namespace Microsoft.Identity.Web
 #endif            
             JwtBearerOptions options)
         {
-            MergedOptions.UpdateMergedOptionsFromJwtBearerOptions(options, _mergedOptionsMonitor.Get(name ?? string.Empty));
+            MergedOptions mergedOptions = _mergedOptionsMonitor.Get(name ?? string.Empty);
+
+            // Take into account extra query parameters
+            UpdateOptionsMetadata(options, mergedOptions);
+            MergedOptions.UpdateMergedOptionsFromJwtBearerOptions(options, mergedOptions);
+        }
+
+        private static void UpdateOptionsMetadata(JwtBearerOptions options, MergedOptions mergedOptions)
+        {
+            options.MetadataAddress ??= options.Authority + "/.well-known/openid-configuration";
+
+            if (mergedOptions.ExtraQueryParameters != null)
+            {
+                options.MetadataAddress += options.MetadataAddress.Contains('?', StringComparison.OrdinalIgnoreCase) ? "&" : "?";
+                options.MetadataAddress += string.Join("&", mergedOptions.ExtraQueryParameters.Select(p => $"{p.Key}={p.Value}"));
+            }
         }
     }
 }
